move sbcc files for ussd to ussd folder
