+++ sbcc/ussd/input-handlers/pinMenuHandler.test.js
@@ -1,9 +1,9 @@
 const pinMenuHandler = require('./pinMenuHandler');
-const notifyELK = require('../../notifications/elk-notification/elkNotification');
-const scheduleCall = require('../utils/scheduleCall');
+const notifyELK = require('../../../notifications/elk-notification/elkNotification');
+const scheduleCall = require('../../utils/scheduleCall');
 
-jest.mock('../../notifications/elk-notification/elkNotification');
-jest.mock('../utils/scheduleCall');
+jest.mock('../../../notifications/elk-notification/elkNotification');
+jest.mock('../../utils/scheduleCall');
 
 describe('Pin Menu Handler', () => {
     beforeAll(() => {
